requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
black==22.6.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
dj-database-url==1.0.0
Django==4.0.6
django-cors-headers==3.13.0
django-environ==0.9.0
django-heroku==0.3.1
django-phonenumber-field==6.3.0
django-rest-passwordreset==1.3.0
djangorestframework==3.13.1
gunicorn==20.1.0
kombu==5.2.4
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
phonenumbers==8.12.52
phonenumberslite==8.12.52
platformdirs==2.5.2
prompt-toolkit==3.0.31
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.4
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
